requirements.txt
Django==1.11.18
line-bot-sdk==1.8.0
